Add files via upload
Implementation in Python of a program to compute jumping numbers of reduced plane curves.
The user inputs the number of relevant exceptional divisors, the number of germs in the maximal contact completion and the number of components of the curve. Further it is required to introduce the orders of vanishing of the elements in the maximal contact completion over the relevant exceptional divisors and their log-discrepancies.
As an output, the program generates the list of jumping numbers <1 together with the monomial generating it and its support (at which divisors the jumping number is attained).